Move code from offline_map_matching to util package
